Split compiled runtime capabilities
