+++ packages/eclipsa/core/runtime/kernel.ts
@@ -1,0 +1,7 @@
+let runtimeSymbols: Record<string, string> = {}
+
+export const setRuntimeSymbols = (symbols: Record<string, string> | undefined) => {
+  runtimeSymbols = symbols ?? {}
+}
+
+export const getRuntimeSymbolUrl = (symbol: string) => runtimeSymbols[symbol]
