+++ packages/eclipsa/package.json
@@ -37,6 +37,26 @@
       "types": "./compiled-client.ts",
       "import": "./compiled-client.ts"
     },
+    "./runtime/dom-compiled": {
+      "types": "./runtime/dom-compiled.ts",
+      "import": "./runtime/dom-compiled.ts"
+    },
+    "./runtime/event": {
+      "types": "./runtime/event.ts",
+      "import": "./runtime/event.ts"
+    },
+    "./runtime/hydrate": {
+      "types": "./runtime/hydrate.ts",
+      "import": "./runtime/hydrate.ts"
+    },
+    "./runtime/reactive": {
+      "types": "./runtime/reactive.ts",
+      "import": "./runtime/reactive.ts"
+    },
+    "./runtime/resume": {
+      "types": "./runtime/resume.ts",
+      "import": "./runtime/resume.ts"
+    },
     "./web-utils": {
       "types": "./web-utils/mod.ts",
       "import": "./web-utils/mod.ts"
@@ -103,6 +123,26 @@
       "./compiled-client": {
         "types": "./compiled-client.d.mts",
         "import": "./compiled-client.mjs"
+      },
+      "./runtime/dom-compiled": {
+        "types": "./runtime/dom-compiled.d.mts",
+        "import": "./runtime/dom-compiled.mjs"
+      },
+      "./runtime/event": {
+        "types": "./runtime/event.d.mts",
+        "import": "./runtime/event.mjs"
+      },
+      "./runtime/hydrate": {
+        "types": "./runtime/hydrate.d.mts",
+        "import": "./runtime/hydrate.mjs"
+      },
+      "./runtime/reactive": {
+        "types": "./runtime/reactive.d.mts",
+        "import": "./runtime/reactive.mjs"
+      },
+      "./runtime/resume": {
+        "types": "./runtime/resume.d.mts",
+        "import": "./runtime/resume.mjs"
       },
       "./web-utils": {
         "types": "./web-utils/mod.d.mts",
