+++ packages/eclipsa/runtime/dom-compiled.ts
@@ -1,0 +1,1 @@
+export * from '../core/runtime/dom-compiled.ts'
