+++ packages/eclipsa/runtime/event.ts
@@ -1,0 +1,1 @@
+export * from '../core/runtime/event.ts'
