+++ packages/eclipsa/runtime/hydrate.ts
@@ -1,0 +1,1 @@
+export * from '../core/runtime/hydrate.ts'
