+++ packages/eclipsa/runtime/resume.ts
@@ -1,0 +1,1 @@
+export { hydrate } from '../core/client/mod.ts'
