+++ packages/eclipsa/vite.config.ts
@@ -18,6 +18,11 @@ export default defineConfig({
       'compiled-client.ts',
       'flow.ts',
       'meta.ts',
+      'runtime/dom-compiled.ts',
+      'runtime/event.ts',
+      'runtime/hydrate.ts',
+      'runtime/reactive.ts',
+      'runtime/resume.ts',
       'signal.ts',
       'web-utils/mod.ts',
       'vite/mod.ts',
